feat(artifacts): route model artifacts through the object store (#22) (#123)
* feat(artifacts): route model artifacts through the object store (#22)

Fine-tune adapters and context-predictor weights now write to and reload
from the configured object store, so a cross-host worker fleet shares
trained models. A local-FS root stays valid for single-host; an S3/MinIO
root lets workers on different hosts read each other's artifacts.

Correctness model: catalog-pointer-as-commit. A worker writes artifact
objects to a unique per-attempt prefix ({job_id}/{worker_id}/{attempt});
the existing lease-guarded finalize CAS records that prefix into
models.artifact_path. The catalog row update is the single atomic commit —
no object is moved or renamed, so there is no torn-promote window. Losing
attempts' prefixes are orphaned and best-effort GC'd.

- jammi-db: new ArtifactStore (sibling to ResultStore) over the shared
  StorageRegistry. put_artifact writes files then a manifest.json LAST
  (exact keys + per-file sha256); fetch_artifact is manifest-driven (never
  LISTs), sha256-verified, into a content-addressed local cache via
  tempdir-download + atomic-rename (torn-free under concurrent fetch). A
  file:// root reads in place (no copy). delete_artifact_prefix GCs an
  orphaned prefix.
- jammi-ai write path: trainer/predictor produce final files in a local
  tempdir; the worker's unified publish_and_finalize writes them under the
  per-attempt prefix and records that prefix in the finalize CAS. Deletes
  StagedArtifact (stage/promote/discard) entirely — the catalog-pointer
  commit replaces it.
- jammi-ai reload path: the context-predictor reload, the fine-tuned
  adapter resolve, and the worker's catalog-artifact_path base-model load
  all parse artifact_path as a StorageUrl and fetch_artifact to a local
  cache dir; the HF-hub fallback is unchanged.
- Root: model artifacts live under a models/ sub-prefix of the existing
  storage.result_root (no new config knob); fetch cache under
  {artifact_dir}/jammi_db/artifact_cache.
- Tests: ArtifactStore contract tests; cross-host proof in storage_root
  (artifact_written_on_host_a_is_loadable_on_host_b — shared object store,
  fresh local cache on host B); existing fine-tune/predictor/encoder/graph
  tests migrated to fetch via the artifact store.


* fix(artifacts): make the lease-guarded finalize CAS the sole writer of the served artifact_path

Under reclaim, a zombie worker (lease expired, still executing its claimed
run) could corrupt a completed model: the served prefix was committed by an
unguarded, last-writer-wins register_model, so the zombie's late register
overwrote models.artifact_path with its own prefix, its CAS-loss branch then
deleted that prefix's bytes, and a separate unguarded run-start status write
regressed the job completed → running. The completed model became permanently
unloadable (manifest.json NotFound).

Fix — catalog-pointer-as-commit, done correctly (shape (a)): the served path is
written only by the worker whose lease-guarded finalize CAS wins, and by no one
else.

- Promote the served artifact_path to a first-class models column (migration
  017). This lets finalize set it with a plain UPDATE in the same transaction
  as the job-row CAS — no dialect-specific JSON mutation — and keeps the reload
  path reading the pointer straight from the model row, unchanged. The
  descriptive base_model_id/config_json stay in the metadata blob.
- finalize_training_job takes the winner's prefix and, in one transaction, runs
  the job-row CAS (claimed_by + status='running') and — only when that CAS
  matched — records artifact_path on the output model's row. A loser's finalize
  matches zero rows and writes nothing (neither status nor served path). Both
  the SQLite and Postgres impls go through this single backend-agnostic path.
- register_model no longer sets the served path for a finalized job: the worker
  registers the output row with artifact_path = None (so a completed observer
  finds a row), and on ON CONFLICT the path is COALESCE(excluded, existing) — a
  None can never clear or overwrite a committed pointer. A base model still sets
  its own file:// path directly; that path is never job-contended.
- CAS-loss branch: the loser never commits a path, so delete_artifact_prefix on
  its own prefix is unconditionally safe (it is never the committed pointer).
- Secondary regression root cause: the trainer's run-start status write was
  unguarded. Replace it with a lease-guarded mark_training_running
  (claimed_by + status='running'); a reclaimed zombie's trainer start now
  matches zero rows and cannot regress a terminal job to running. The trainer
  takes the lease holder's worker_id for the guard.
- Fix the false doc comment that claimed the path is "overwritten by whichever
  worker's CAS wins"; describe the real single-writer invariant.

Regression test (winner-first): the winner trains, CAS-wins, completes; THEN
the stale loser runs to completion and registers/finalizes. Asserts the served
path is the winner's prefix, those bytes still fetch (reload works), and the
job stays completed. Fails pre-fix (committed path is the zombie's,
manifest.json NotFound), passes post-fix. The db-level
finalize_is_a_lease_guarded_compare_and_set test gains served-path assertions:
the loser's failed CAS commits no path (stays NULL), the winner's commits its
prefix.


* fix(catalog): decode INTEGER columns as i32 for Postgres INT4; drop dead unguarded status writer

Fix 1 — Postgres INT4 decode mismatch (CI failure)
Every `INTEGER` column is `INT4` (32-bit) on Postgres; sqlx-Postgres is strict
and rejects an `i64` (`INT8`) decode with `ColumnDecode`, while SQLite's loose
integer typing hid it. `models.version` was decoded as `i64`, panicking the
`Test (Postgres)` lane at model-row decode (the same class as the prior
`attempts` i32 fix). Decode the whole class as `i32`, matching the schema
column width — portable across both backends:
- models.version          (model_repo.rs)
- result_tables.row_count, .checkpoint (result_repo.rs)
- eval_runs.k             (eval_repo.rs)
- evidence_channels.priority, the `SELECT 1` existence probe, and the
  `MAX(ordinal)` aggregate (channel_repo.rs)
`is_unique` stays `i64` — its column is `BIGINT` by design; `COUNT(*)`
aggregates stay `i64` (Postgres returns INT8). Binds keep i64 params against
INT4 columns (Postgres widens on assignment/comparison), so no bind change.

Fix 2 — delete the dead, unguarded status writer
`update_training_status` had no production callers (the trainer uses the
lease-guarded `mark_training_running`) and was the only remaining unguarded
`UPDATE training_jobs SET status … WHERE job_id` — a latent reclaim footgun.
Deleted. Retargeted its two test usages to the guarded API:
- jammi-db heartbeat test: drive the job out of `running` via
  `fail_training_job` (the lease-guarded terminal transition).
- jammi-ai lifecycle test: stamp run-start metrics via
  `mark_training_running` (the lease-guarded non-terminal write).

Fix 3 — migration 017 doc comment
Rewrote the journey-shaped narration (what the path "used to be" inside the
metadata blob) to describe what the `artifact_path` column IS: the
single-writer commit pointer set by the lease-guarded finalize CAS.


---------

+++ crates/jammi-db/src/catalog/result_repo.rs
@@ -105,7 +105,7 @@ fn parse_row(row: &Row<'_>) -> std::result::Result<ResultTableRecord, BackendErr
         index_path: row.try_get("index_path")?,
         dimensions: row.try_get("dimensions")?,
         distance_metric: row.get("distance_metric")?,
-        row_count: row.get::<i64>("row_count")? as usize,
+        row_count: row.get::<i32>("row_count")? as usize,
         status: row.get("status")?,
         key_column: row.try_get("key_column")?,
         text_columns: row.try_get("text_columns")?,
@@ -451,7 +451,7 @@ impl Catalog {
     pub async fn get_checkpoint(&self, name: &str) -> Result<Option<usize>> {
         let name = name.to_string();
         let tenant = self.current_tenant();
-        let found: Option<Option<i64>> = self
+        let found: Option<Option<i32>> = self
             .backend()
             .transaction(
                 TxOptions {
@@ -467,7 +467,7 @@ impl Catalog {
                                 SqlValue::TextOwned(name),
                                 SqlValue::from(tenant.map(|t| t.to_string())),
                             ],
-                            |row| row.try_get::<i64>("checkpoint"),
+                            |row| row.try_get::<i32>("checkpoint"),
                         )
                         .await
                     })
